backend/requirements.txt
django==5.0.*
djangorestframework==3.15.*
django-cors-headers==4.3.*
django-allauth==0.62.*
psycopg2-binary==2.9.*
python-decouple==3.8.*
gunicorn==21.2.*
dj-database-url==2.1.*
